issue 55: hidden creation tools if window is too small moving all tool groups to root palette updating GmfToolTestCase to match new gmftool structure

+++ trunk/org.openiaml.model/model/scripts/apply.bat
@@ -3,5 +3,5 @@ rem -- apply the diff for %1 model .gmfgen --
 
 cd ..
 patch %1.gmfgen diff/%1.gmfgen.diff
-rm *.orig
+if exist *.orig rm *.orig
 cd scripts
